# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jordan's 40th Trip Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /households:
    get:
      operationId: listHouseholds
      tags: [households]
      summary: List all households
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Household"
    post:
      operationId: createHousehold
      tags: [households]
      summary: Create a household
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseholdInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
  /households/{id}:
    patch:
      operationId: updateHousehold
      tags: [households]
      summary: Update a household
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseholdUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
    delete:
      operationId: deleteHousehold
      tags: [households]
      summary: Delete a household
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ---- Guests ----
  /guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List all guests
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
    post:
      operationId: createGuest
      tags: [guests]
      summary: Create a guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{id}:
    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    delete:
      operationId: deleteGuest
      tags: [guests]
      summary: Delete a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ---- Rooms ----
...
  /rooms/{id}:
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

  # ---- Expenses ----
...
  /itinerary:
    get:
      operationId: listItineraryItems
      tags: [itinerary]
      summary: List all itinerary items
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItineraryItem"
    post:
      operationId: createItineraryItem
      tags: [itinerary]
      summary: Create an itinerary item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItineraryItemInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItineraryItem"
  /itinerary/{id}:
    patch:
      operationId: updateItineraryItem
      tags: [itinerary]
      summary: Update an itinerary item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItineraryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItineraryItem"
    delete:
      operationId: deleteItineraryItem
      tags: [itinerary]
      summary: Delete an itinerary item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    # ---- Households ----
    Household:
      type: object
      properties:
        id:        { type: number }
        name:      { type: string }
        photoUrl:  { type: ["string", "null"] }
        phone:     { type: ["string", "null"] }
        email:     { type: ["string", "null"] }
        notes:     { type: ["string", "null"] }
        sortOrder: { type: ["number", "null"] }
      required: [id, name]
    HouseholdInput:
      type: object
      properties:
        name:      { type: string }
        photoUrl:  { type: ["string", "null"] }
        phone:     { type: ["string", "null"] }
        email:     { type: ["string", "null"] }
        notes:     { type: ["string", "null"] }
        sortOrder: { type: number }
      required: [name]
    HouseholdUpdate:
      type: object
      properties:
        name:      { type: string }
        photoUrl:  { type: ["string", "null"] }
        phone:     { type: ["string", "null"] }
        email:     { type: ["string", "null"] }
        notes:     { type: ["string", "null"] }
        sortOrder: { type: number }

    # ---- Guests ----
    Guest:
      type: object
      properties:
        id:                   { type: number }
        householdId:          { type: ["number", "null"] }
        name:                 { type: string }
        arrivalDatetime:      { type: ["string", "null"] }
        airline:              { type: ["string", "null"] }
        arrivalFlightNumber:  { type: ["string", "null"] }
        arrivalAirport:       { type: ["string", "null"] }
        departureDatetime:    { type: ["string", "null"] }
        departureFlightNumber: { type: ["string", "null"] }
        departureAirport:     { type: ["string", "null"] }
        rentalCarInfo:        { type: ["string", "null"] }
        transportationNeeds:  { type: ["string", "null"] }
        flightStatus:         { type: ["string", "null"] }
        notes:                { type: ["string", "null"] }
      required: [id, name]
    GuestInput:
      type: object
      properties:
        householdId:          { type: ["number", "null"] }
        name:                 { type: string }
        arrivalDatetime:      { type: ["string", "null"] }
        airline:              { type: ["string", "null"] }
        arrivalFlightNumber:  { type: ["string", "null"] }
        arrivalAirport:       { type: ["string", "null"] }
        departureDatetime:    { type: ["string", "null"] }
        departureFlightNumber: { type: ["string", "null"] }
        departureAirport:     { type: ["string", "null"] }
        rentalCarInfo:        { type: ["string", "null"] }
        transportationNeeds:  { type: ["string", "null"] }
        flightStatus:         { type: string }
        notes:                { type: ["string", "null"] }
      required: [name]
    GuestUpdate:
      type: object
      properties:
        householdId:          { type: ["number", "null"] }
        name:                 { type: string }
        arrivalDatetime:      { type: ["string", "null"] }
        airline:              { type: ["string", "null"] }
        arrivalFlightNumber:  { type: ["string", "null"] }
        arrivalAirport:       { type: ["string", "null"] }
        departureDatetime:    { type: ["string", "null"] }
        departureFlightNumber: { type: ["string", "null"] }
        departureAirport:     { type: ["string", "null"] }
        rentalCarInfo:        { type: ["string", "null"] }
        transportationNeeds:  { type: ["string", "null"] }
        flightStatus:         { type: string }
        notes:                { type: ["string", "null"] }

    # ---- Rooms ----
    Room:
      type: object
      properties:
        id:                   { type: number }
        name:                 { type: string }
        bedConfig:            { type: ["string", "null"] }
        bathLayout:           { type: ["string", "null"] }
        occupancy:            { type: ["number", "null"] }
        photoUrl:             { type: ["string", "null"] }
        notes:                { type: ["string", "null"] }
        assignedHouseholdId:  { type: ["number", "null"] }
        sortOrder:            { type: ["number", "null"] }
      required: [id, name]
    RoomUpdate:
      type: object
      properties:
        name:                 { type: string }
        bedConfig:            { type: ["string", "null"] }
        bathLayout:           { type: ["string", "null"] }
        occupancy:            { type: ["number", "null"] }
        photoUrl:             { type: ["string", "null"] }
        notes:                { type: ["string", "null"] }
        assignedHouseholdId:  { type: ["number", "null"] }
        sortOrder:            { type: number }

    # ---- Expenses ----
...
    ItineraryItem:
      type: object
      properties:
        id:          { type: number }
        dayLabel:    { type: string }
        time:        { type: ["string", "null"] }
        title:       { type: string }
        description: { type: ["string", "null"] }
        category:    { type: ["string", "null"] }
        sortOrder:   { type: ["number", "null"] }
      required: [id, dayLabel, title]
    ItineraryItemInput:
      type: object
      properties:
        dayLabel:    { type: string }
        time:        { type: ["string", "null"] }
        title:       { type: string }
        description: { type: ["string", "null"] }
        category:    { type: string }
        sortOrder:   { type: number }
      required: [dayLabel, title]
    ItineraryItemUpdate:
      type: object
      properties:
        dayLabel:    { type: string }
        time:        { type: ["string", "null"] }
        title:       { type: string }
        description: { type: ["string", "null"] }
        category:    { type: string }
        sortOrder:   { type: number }