# lib/api-zod/src/generated/types/guestUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jordan's 40th Trip Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface GuestUpdate {
  /** @nullable */
  householdId?: number | null;
  name?: string;
  /** @nullable */
  arrivalDatetime?: string | null;
  /** @nullable */
  airline?: string | null;
  /** @nullable */
  arrivalFlightNumber?: string | null;
  /** @nullable */
  arrivalAirport?: string | null;
  /** @nullable */
  departureDatetime?: string | null;
  /** @nullable */
  departureFlightNumber?: string | null;
  /** @nullable */
  departureAirport?: string | null;
  /** @nullable */
  rentalCarInfo?: string | null;
  /** @nullable */
  transportationNeeds?: string | null;
  flightStatus?: string;
  /** @nullable */
  notes?: string | null;
}