# lib/api-zod/src/generated/types/householdUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jordan's 40th Trip Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface HouseholdUpdate {
  name?: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
  sortOrder?: number;
}