# lib/api-zod/src/generated/types/itineraryItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jordan's 40th Trip Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ItineraryItem {
  id: number;
  dayLabel: string;
  /** @nullable */
  time?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sortOrder?: number | null;
}