# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jordan's 40th Trip Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  bedConfig?: string | null;
  /** @nullable */
  bathLayout?: string | null;
  /** @nullable */
  occupancy?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  assignedHouseholdId?: number | null;
  /** @nullable */
  sortOrder?: number | null;
}